move utils into utils folder
