+++ src/aframe-mods/aabb-collider.js
@@ -1,6 +1,6 @@
 /* global AFRAME, THREE */
 
-import {ARENAUtils} from '../utils.js';
+import {ARENAUtils} from '../utils';
 
 if (typeof AFRAME === 'undefined') {
     throw new Error('Component attempted to register before AFRAME was available.');
