+++ src/utils/index.js
@@ -1,0 +1,7 @@
+import { ARENAUtils } from './utils';
+import { ARENAMqttConsole } from './arena-console';
+
+export {
+    ARENAUtils,
+    ARENAMqttConsole
+}
